backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
defusedxml==0.7.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
httpie==3.2.4
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
PySocks==1.7.1
python-dotenv==1.2.2
python-json-logger==4.0.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
